% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{pkg_template}
\alias{pkg_template}
\title{Start a new project with the cpp4r package template}
\usage{
pkg_template(path = NULL, pkgname = NULL)
}
\arguments{
\item{path}{Path to the new project}

\item{pkgname}{Name of the new package}
}
\value{
The file path to the copied template (invisibly).
}
\description{
Copies a package template into a new directory.
The template includes a DESCRIPTION file, a minimal R/ directory and placeholders
with instructions. You can then edit these files to customize your new package.
}
\examples{
# create a new directory
dir <- tempdir()
dir.create(dir)

# copy the package template into the directory
pkg_template(dir, "mynewpkg")
}
